[NEW] - Content/ Website [ADD] Favicon, Second Page
In the following update the website instead of using default css and html now use " just the docs " basic template for its website design that is customized and updated for readability and usability.

New Content, Getting Started :
Accessing or getting access to CSS (IN CONStruction)
HTML AND CSS use cases

Home Page, and additionally a link to the repository.