[Fix again] - Hardcoded Links to pages...
Since it refuses to find the next page from itself, it can only be used as a full direct hyperlink to the page I want it to go to.

Nothing else that can be done about it, as the API for the following is broken specifically for the getting started page.